requirements.txt
alembic==1.17.2
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1754315087/work
boto3==1.42.4
botocore==1.42.4
cachetools==6.2.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
charset-normalizer==3.4.4
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1733256735222/work
ecdsa==0.19.1
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1733300719943/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-core_1752617909/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1751972066250/work
google-auth==2.43.0
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1745526374115/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httptools @ file:///Users/runner/miniforge3/conda-bld/httptools_1762504101475/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jmespath==1.0.1
Mako==1.3.10
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1759055293570/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
pillow==12.0.0
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1749927112905/work
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pydantic-core_1746625311/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
python-jose==3.5.0
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1734420773152/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1758891758214/work
requests==2.32.5
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1753436991/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1755001296/work
rsa==4.9.1
s3transfer==0.16.0
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1733300899265/work
six==1.17.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1753374988/work
tomli==2.3.0
typer==0.16.1
typer-slim==0.16.1
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1747870647094/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
urllib3==2.6.0
uvicorn @ file:///home/conda/feedstock_root/build_artifacts/uvicorn_1751201600733/work
uvloop @ file:///Users/runner/miniforge3/conda-bld/uvloop_1762472822641/work
watchfiles @ file:///Users/runner/miniforge3/conda-bld/watchfiles_1760456881261/work
websockets @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_websockets_1756476464/work/dist
